refactor cloud call tool telemetry
