+++ src/server/routers/tools/index.ts
@@ -1,14 +1,14 @@
 import { publicProcedure, router } from '@/libs/trpc/lambda';
 
-import { codeInterpreterRouter } from './codeInterpreter';
 import { klavisRouter } from './klavis';
+import { marketRouter } from './market';
 import { mcpRouter } from './mcp';
 import { searchRouter } from './search';
 
 export const toolsRouter = router({
-  codeInterpreter: codeInterpreterRouter,
   healthcheck: publicProcedure.query(() => "i'm live!"),
   klavis: klavisRouter,
+  market: marketRouter,
   mcp: mcpRouter,
   search: searchRouter,
 });
